# src/app/api/openapi.yaml
openapi: 3.0.3
info:
  title: Dupip Morpheus API
  version: 1.0.0
  description: >
    HTTP API for the Dupip application. Authentication uses Clerk session
    cookies; most routes require an authenticated user. Errors use the shape
    `{ "error": "message" }`.
servers:
  - url: https://www.dupip.com
    description: Production
  - url: https://beta.dupip.com
    description: Nightly
  - url: http://localhost:3000
    description: Local development
security:
  - cookieAuth: []
...
paths:
...
  /api/v1/tasks:
    get:
      tags: [tasks]
      summary: List tasks for a date (date-aware, RRULE-based)
      parameters:
        - { name: listId, in: query, required: true, schema: { type: string } }
        - { name: date, in: query, required: true, schema: { type: string } }
      responses:
        '200': { description: Tasks }
    post:
      tags: [tasks]
      summary: Create a task (name, listId, rrule, dtstart, times, premium, premiumType, location)
      responses:
        '200': { description: Task }
  /api/v1/tasks/{taskId}:
    get:
      tags: [tasks]
      summary: Get a task (members)
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Task }
    put:
      tags: [tasks]
      summary: Update a task (OWNER/MANAGER/COLLABORATOR; status COMPLETED completes the task)
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Updated }
    delete:
      tags: [tasks]
      summary: Delete a task with scope (OWNER/MANAGER)
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: string } }
        - { name: scope, in: query, schema: { type: string, enum: [all, onwards, today] } }
        - { name: date, in: query, schema: { type: string } }
      responses:
        '200': { description: Deleted }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __session